Update J.O.U.R.N.A.L.py
specified encoding as utf-8 for all open() just in case, hopefuly that didn't break anything. also edited readEntryFile to only require the entry as input, and made user optional for possible future updates, currently unused, since if not set it defaults to currentUser. fixed a bug where the writeEntryFile function would return a error when the user folder is empty. changed the successful import iteration variable to iter, to match the export code. added new entries to dialogue class, changed the names of some of the members of the colors class. made colors.get() return the color white if it can't find the input in the class. changed fileStatus function to not require user in input, and simplified output a bit by removing the isReadable variable. modified the listing system of the entries list to write in a different color the name of the file, and overhauled a lot of the code. finished integration of the new file system, although it hasn't been thoroughly tested yet. removed temporary error messages. might have missed a few smaller changes.